Finalizando proyecto 3 again

+++ laboratorio1/app/src/main/java/com/example/laboratorio1/Calculadora.java
@@ -1,19 +1,19 @@
 package com.example.laboratorio1;
 
 public class Calculadora {
-    public static int suma(int n, int j){
+    public static double suma(double n, double j){
         return n + j;
     }
 
-    public static int resta(int n, int j){
+    public static double resta(double n, double j){
         return n - j;
     }
 
-    public static int multiplicacion(int n, int j){
+    public static double multiplicacion(double n, double j){
         return n * j;
     }
 
-    public static int division(int n, int j){
+    public static double division(double n, double j){
         return n / j;
     }
 }
